requirements.txt
gpt_index==0.8.2.post1
langchain==0.0.283
llama_index==0.8.19
openai==0.28.0
pinecone_client==2.2.2
sentence_transformers==2.2.2
snowflake_connector_python==3.0.3
snowflake_snowpark_python==1.5.0
streamlit==1.26.0
streamlit_chat==0.1.1
streamlit_modal==0.1.0
